Consolidate 'DatasetKind' a bit
